+++ packages/language-support/src/autocompletion/autocompletion.ts
@@ -1,15 +1,17 @@
 import { CompletionItem } from 'vscode-languageserver-types';
 
 import { DbSchema } from '../dbSchema';
-import { findLatestStatement } from '../helpers';
+import { findCaret } from '../helpers';
 import { parserWrapper } from '../parserWrapper';
 import { completionCoreCompletion } from './completionCoreCompletions';
 
 export function autocomplete(
-  textUntilPosition: string,
+  query: string,
   dbSchema: DbSchema,
+  caretPosition: number = query.length,
+  manual = false,
 ): CompletionItem[] {
-  let parsingResult = parserWrapper.parse(textUntilPosition);
+  const parsingResult = parserWrapper.parse(query);
   /* We try to locate the latest statement by finding the latest available `;` 
      in the query and take from that point to the end of the query
 
@@ -22,11 +24,11 @@ export function autocomplete(
      If there was no ;, we don't want to reparse, so we return undefined 
      inside findLatestStatement
   */
-
-  const lastStatement = findLatestStatement(parsingResult);
-  if (lastStatement != undefined) {
-    parsingResult = parserWrapper.parse(lastStatement);
+  const caret = findCaret(parsingResult, caretPosition);
+  if (caret) {
+    const statement = caret.statement;
+    return completionCoreCompletion(statement, dbSchema, manual);
   }
 
-  return completionCoreCompletion(parsingResult, dbSchema);
+  return [];
 }