+++ packages/language-support/src/autocompletion/completionCoreCompletions.ts
@@ -4,24 +4,24 @@ import { CodeCompletionCore } from 'antlr4-c3';
 import {
   CompletionItem,
   CompletionItemKind,
+  InsertTextFormat,
 } from 'vscode-languageserver-types';
 import { DbSchema } from '../dbSchema';
 import CypherLexer from '../generated-parser/CypherCmdLexer';
 import CypherParser, {
   Expression2Context,
 } from '../generated-parser/CypherCmdParser';
-import { rulesDefiningVariables } from '../helpers';
+import { findPreviousNonSpace, rulesDefiningVariables } from '../helpers';
 import {
   CypherTokenType,
   lexerKeywords,
   lexerSymbols,
   tokenNames,
 } from '../lexerSymbols';
-import {
-  consoleCommandEnabled,
-  EnrichedParsingResult,
-  ParsingResult,
-} from '../parserWrapper';
+
+import { ParsedStatement } from '../parserWrapper';
+
+import { consoleCommandEnabled } from '../parserWrapper';
 
 const uniq = <T>(arr: T[]) => Array.from(new Set(arr));
 
@@ -45,7 +45,7 @@ const functionNameCompletions = (
   namespacedCompletion(
     candidateRule,
     tokens,
-    Object.keys(dbSchema?.functionSignatures ?? {}),
+    Object.keys(dbSchema?.functions ?? {}),
     'function',
   );
 
@@ -57,7 +57,7 @@ const procedureNameCompletions = (
   namespacedCompletion(
     candidateRule,
     tokens,
-    Object.keys(dbSchema?.procedureSignatures ?? {}),
+    Object.keys(dbSchema?.procedures ?? {}),
     'procedure',
   );
 
@@ -216,12 +216,27 @@ enum ExpectedParameterType {
 
 const inferExpectedParameterTypeFromContext = (context: CandidateRule) => {
   const parentRule = context.ruleList.at(-1);
+  const grandParentRule = context.ruleList.at(-2);
   if (
     [
       CypherParser.RULE_stringOrParameter,
       CypherParser.RULE_symbolicNameOrStringParameter,
+      CypherParser.RULE_symbolicNameOrStringParameterList,
+      CypherParser.RULE_symbolicAliasNameOrParameter,
       CypherParser.RULE_passwordExpression,
-    ].includes(parentRule)
+      CypherParser.RULE_createUser,
+      CypherParser.RULE_dropUser,
+      CypherParser.RULE_alterUser,
+      CypherParser.RULE_renameUser,
+      CypherParser.RULE_createRole,
+      CypherParser.RULE_dropRole,
+      CypherParser.RULE_renameRole,
+      CypherParser.RULE_revokeRole,
+    ].includes(parentRule) ||
+    [
+      CypherParser.RULE_showUserPrivileges,
+      CypherParser.RULE_grantRole,
+    ].includes(grandParentRule)
   ) {
     return ExpectedParameterType.String;
   } else if (
@@ -284,8 +299,9 @@ function calculateNamespacePrefix(
 }
 
 export function completionCoreCompletion(
-  parsingResult: EnrichedParsingResult,
+  parsingResult: ParsedStatement,
   dbSchema: DbSchema,
+  manualTrigger = false,
 ): CompletionItem[] {
   const parser = parsingResult.parser;
   const tokens = parsingResult.tokens;
@@ -305,10 +321,10 @@ export function completionCoreCompletion(
   }
 
   // If the previous token is an identifier, we don't count it as "finished" so we move the caret back one token
-  // The identfier is finished when the last token is a SPACE or dot etc. etc.
+  // The identifier is finished when the last token is a SPACE or dot etc. etc.
   // this allows us to give completions that replace the current text => for example `RET` <- it's parsed as an identifier
   // The need for this caret movement is outlined in the documentation of antlr4-c3 in the section about caret position
-  // When an identifer overlaps with a keyword, it's no longer treates as an identifier (although it's a valid identifier)
+  // When an identifier overlaps with a keyword, it's no longer treats as an identifier (although it's a valid identifier)
   // So we need to move the caret back for keywords as well
   const previousToken = tokens[caretIndex - 1]?.type;
   if (
@@ -326,6 +342,9 @@ export function completionCoreCompletion(
     CypherParser.RULE_parameter,
     CypherParser.RULE_propertyKeyName,
     CypherParser.RULE_variable,
+    CypherParser.RULE_leftArrow,
+    // this rule is used for usernames and roles.
+    CypherParser.RULE_symbolicNameOrStringParameter,
 
     // Either enable the helper rules for lexer clashes,
     // or collect all console commands like below with symbolicNameString
@@ -383,7 +402,7 @@ export function completionCoreCompletion(
         const parentRule = candidateRule.ruleList.at(-1);
         const grandParentRule = candidateRule.ruleList.at(-2);
         if (
-          parentRule === CypherParser.RULE_mapLiteral &&
+          parentRule === CypherParser.RULE_map &&
           grandParentRule === CypherParser.RULE_literal
         ) {
           return [];
@@ -397,7 +416,7 @@ export function completionCoreCompletion(
         // but it is likely to be a node/relationship
         if (
           parentRule === CypherParser.RULE_property &&
-          grandParentRule == CypherParser.RULE_postFix1 &&
+          grandParentRule == CypherParser.RULE_postFix &&
           greatGrandParentRule === CypherParser.RULE_expression2
         ) {
           const expr2 = parsingResult.stopNode?.parentCtx?.parentCtx?.parentCtx;
@@ -435,6 +454,10 @@ export function completionCoreCompletion(
         return completeAliasName({ candidateRule, dbSchema, parsingResult });
       }
 
+      if (ruleNumber === CypherParser.RULE_symbolicNameOrStringParameter) {
+        return completeSymbolicName({ candidateRule, dbSchema, parsingResult });
+      }
+
       if (ruleNumber === CypherParser.RULE_labelExpression1) {
         const topExprIndex = candidateRule.ruleList.indexOf(
           CypherParser.RULE_labelExpression,
@@ -454,12 +477,7 @@ export function completionCoreCompletion(
           return reltypeCompletions(dbSchema);
         }
 
-        if (topExprParent === CypherParser.RULE_labelExpressionPredicate) {
-          return [
-            ...labelCompletions(dbSchema),
-            ...reltypeCompletions(dbSchema),
-          ];
-        }
+        return [...labelCompletions(dbSchema), ...reltypeCompletions(dbSchema)];
       }
 
       // These are simple tokens that get completed as the wrong kind, due to a lexer conflict
@@ -471,9 +489,49 @@ export function completionCoreCompletion(
         return [{ label: 'list', kind: CompletionItemKind.Event }];
       }
 
+      if (ruleNumber === CypherParser.RULE_leftArrow) {
+        return [
+          {
+            label: '-[]->()',
+            kind: CompletionItemKind.Snippet,
+            insertTextFormat: InsertTextFormat.Snippet,
+            insertText: '-[${1: }]->(${2: })',
+            detail: 'path template',
+            // vscode does not call the completion provider for every single character
+            // after the second character is typed (i.e) `MATCH ()-[` the completion is no longer valid
+            // it'd insert `MATCH ()-[[]->()` which is not valid. Hence we filter it out by using the filterText
+            filterText: '',
+          },
+          {
+            label: '-[]-()',
+            kind: CompletionItemKind.Snippet,
+            insertTextFormat: InsertTextFormat.Snippet,
+            insertText: '-[${1: }]-(${2: })',
+            detail: 'path template',
+            filterText: '',
+          },
+          {
+            label: '<-[]-()',
+            kind: CompletionItemKind.Snippet,
+            insertTextFormat: InsertTextFormat.Snippet,
+            insertText: '<-[${1: }]-(${2: })',
+            detail: 'path template',
+            filterText: '',
+          },
+        ];
+      }
+
       return [];
     },
   );
+
+  // if the completion was automatically triggered by a snippet trigger character
+  // we should only return snippet completions
+  if (CypherLexer.RPAREN === previousToken && !manualTrigger) {
+    return ruleCompletions.filter(
+      (completion) => completion.kind === CompletionItemKind.Snippet,
+    );
+  }
 
   return [
     ...ruleCompletions,
@@ -482,7 +540,7 @@ export function completionCoreCompletion(
 }
 
 type CompletionHelperArgs = {
-  parsingResult: ParsingResult;
+  parsingResult: ParsedStatement;
   dbSchema: DbSchema;
   candidateRule: CandidateRule;
 };
@@ -554,4 +612,85 @@ function completeAliasName({
         kind: CompletionItemKind.Value,
       })),
   ];
+}
+
+function completeSymbolicName({
+  candidateRule,
+  dbSchema,
+  parsingResult,
+}: CompletionHelperArgs): CompletionItem[] {
+  // parameters are valid values in all cases of symbolic name
+  const baseSuggestions = parameterCompletions(
+    dbSchema,
+    inferExpectedParameterTypeFromContext(candidateRule),
+  );
+
+  const rulesCreatingNewUserOrRole = [
+    CypherParser.RULE_createUser,
+    CypherParser.RULE_createRole,
+  ];
+
+  const previousToken = findPreviousNonSpace(
+    parsingResult.tokens,
+    candidateRule.startTokenIndex,
+  );
+  const afterToToken = previousToken.type === CypherParser.TO;
+  const ruleList = candidateRule.ruleList;
+
+  // avoid suggesting existing user names or role names when creating a new one
+  if (
+    rulesCreatingNewUserOrRole.some((rule) => ruleList.includes(rule)) ||
+    // We are suggesting an user as target for the renaming
+    //      RENAME USER existing TO target
+    // so target should be non-existent
+    (ruleList.includes(CypherParser.RULE_renameUser) && afterToToken)
+  ) {
+    return baseSuggestions;
+  }
+
+  const rulesThatAcceptExistingUsers = [
+    CypherParser.RULE_dropUser,
+    CypherParser.RULE_renameUser,
+    CypherParser.RULE_alterUser,
+    CypherParser.RULE_showUserPrivileges,
+    CypherParser.RULE_roleUser,
+    CypherParser.RULE_showPrivilege,
+    CypherParser.RULE_dbmsPrivilege,
+    CypherParser.RULE_databasePrivilege,
+  ];
+
+  if (rulesThatAcceptExistingUsers.some((rule) => ruleList.includes(rule))) {
+    return [
+      ...baseSuggestions,
+      ...(dbSchema?.userNames ?? []).map((userName) => ({
+        label: userName,
+        kind: CompletionItemKind.Value,
+      })),
+    ];
+  }
+
+  const rulesThatAcceptExistingRoles = [
+    CypherParser.RULE_grantRole,
+    CypherParser.RULE_revokeRole,
+    CypherParser.RULE_dropRole,
+    CypherParser.RULE_renameRole,
+    CypherParser.RULE_showRolePrivileges,
+    CypherParser.RULE_grantRoleManagement,
+    CypherParser.RULE_revokeRoleManagement,
+    CypherParser.RULE_grantPrivilege,
+    CypherParser.RULE_denyPrivilege,
+    CypherParser.RULE_revokePrivilege,
+  ];
+
+  if (rulesThatAcceptExistingRoles.some((rule) => ruleList.includes(rule))) {
+    return [
+      ...baseSuggestions,
+      ...(dbSchema?.roleNames ?? []).map((roleName) => ({
+        label: roleName,
+        kind: CompletionItemKind.Value,
+      })),
+    ];
+  }
+
+  return [];
 }