+++ packages/language-support/src/dbSchema.ts
@@ -1,12 +1,14 @@
-import { SignatureInformation } from 'vscode-languageserver-types';
+import { Neo4jFunction, Neo4jProcedure } from './types';
 
 export interface DbSchema {
-  procedureSignatures?: Record<string, SignatureInformation>;
-  functionSignatures?: Record<string, SignatureInformation>;
   labels?: string[];
   relationshipTypes?: string[];
   databaseNames?: string[];
   aliasNames?: string[];
+  userNames?: string[];
+  roleNames?: string[];
   parameters?: Record<string, unknown>;
   propertyKeys?: string[];
+  procedures?: Record<string, Neo4jProcedure>;
+  functions?: Record<string, Neo4jFunction>;
 }