+++ packages/language-support/src/syntaxColouring/syntaxColouring.ts
@@ -1,8 +1,6 @@
 import { ParseTreeWalker, TerminalNode, Token } from 'antlr4';
 
 import {
-  AllExpressionContext,
-  AnyExpressionContext,
   ArrowLineContext,
   BooleanLiteralContext,
   ConsoleCommandContext,
@@ -12,7 +10,7 @@ import {
   LabelNameIsContext,
   LabelOrRelTypeContext,
   LeftArrowContext,
-  NoneExpressionContext,
+  ListItemsPredicateContext,
   NumberLiteralContext,
   ParameterContext,
   ParamsArgsContext,
@@ -21,9 +19,8 @@ import {
   PropertyKeyNameContext,
   ReduceExpressionContext,
   RightArrowContext,
-  SingleExpressionContext,
+  StringLiteralContext,
   StringsLiteralContext,
-  StringTokenContext,
   SymbolicNameStringContext,
   UseCompletionRuleContext,
   VariableContext,
@@ -178,7 +175,8 @@ class SyntaxHighlighter extends CypherParserListener {
     this.addToken(ctx.start, CypherTokenType.property, ctx.getText());
   };
 
-  exitStringToken = (ctx: StringTokenContext) => {
+  // TODO Do we need this one and the one below?
+  exitStringLiteral = (ctx: StringLiteralContext) => {
     this.addToken(ctx.start, CypherTokenType.stringLiteral, ctx.getText());
   };
 
@@ -210,20 +208,11 @@ class SyntaxHighlighter extends CypherParserListener {
     );
   };
 
-  exitAllExpression = (ctx: AllExpressionContext) => {
-    this.colourPredicateFunction(ctx.ALL());
-  };
-
-  exitAnyExpression = (ctx: AnyExpressionContext) => {
-    this.colourPredicateFunction(ctx.ANY());
-  };
-
-  exitNoneExpression = (ctx: NoneExpressionContext) => {
-    this.colourPredicateFunction(ctx.NONE());
-  };
-
-  exitSingleExpression = (ctx: SingleExpressionContext) => {
-    this.colourPredicateFunction(ctx.SINGLE());
+  exitListItemsPredicate = (ctx: ListItemsPredicateContext) => {
+    if (ctx.ANY()) this.colourPredicateFunction(ctx.ANY());
+    if (ctx.ALL()) this.colourPredicateFunction(ctx.ALL());
+    if (ctx.NONE()) this.colourPredicateFunction(ctx.NONE());
+    if (ctx.SINGLE()) this.colourPredicateFunction(ctx.SINGLE());
   };
 
   exitReduceExpression = (ctx: ReduceExpressionContext) => {
@@ -307,12 +296,7 @@ export function applySyntaxColouring(
   wholeFileText: string,
 ): ParsedCypherToken[] {
   const parsingResult = parserWrapper.parse(wholeFileText);
-  const tokens = parsingResult.tokens;
-
-  // Get a first pass at the colouring using only the lexer
-  const lexerTokens: Map<string, ParsedCypherToken> = colourLexerTokens(tokens);
-
-  const treeSyntaxHighlighter = new SyntaxHighlighter(lexerTokens);
+  const statements = parsingResult.statementsParsing;
 
   /* Get a second pass at the colouring correcting the colours
      using structural information from the parsing tree
@@ -321,9 +305,21 @@ export function applySyntaxColouring(
      recognized as keywords by the lexer in positions
      where they are not keywords (e.g. MATCH (MATCH: MATCH))
   */
-  ParseTreeWalker.DEFAULT.walk(treeSyntaxHighlighter, parsingResult.result);
+  const allColouredTokens: Map<string, ParsedCypherToken> = new Map();
+  statements.forEach((statement) => {
+    const tokens = statement.tokens;
 
-  const allColouredTokens = treeSyntaxHighlighter.colouredTokens;
+    // Get a first pass at the colouring using only the lexer
+    const lexerTokens: Map<string, ParsedCypherToken> =
+      colourLexerTokens(tokens);
+    const treeSyntaxHighlighter = new SyntaxHighlighter(lexerTokens);
+
+    ParseTreeWalker.DEFAULT.walk(treeSyntaxHighlighter, statement.ctx);
+
+    treeSyntaxHighlighter.colouredTokens.forEach((value, key) =>
+      allColouredTokens.set(key, value),
+    );
+  });
 
   // When we push to the builder, tokens need to be sorted in ascending
   // starting position i.e. as we find them when we read them from left