+++ packages/vscode-extension/e2e_tests/helpers.ts
@@ -12,6 +12,16 @@ export async function openDocument(docUri: Uri) {
   }
 }
 
+export async function newUntitledFileWithContent(content: string) {
+  try {
+    // The language server will not be activated automatically
+    const document = await workspace.openTextDocument({ content: content });
+    await window.showTextDocument(document);
+  } catch (e) {
+    console.error(e);
+  }
+}
+
 async function sleep(ms: number) {
   return new Promise((resolve) => setTimeout(resolve, ms));
 }
@@ -22,7 +32,7 @@ export function getDocumentUri(docName: string) {
 
 export async function eventually(
   assertion: () => Promise<void>,
-  timeoutMs = 10000,
+  timeoutMs = 1000000,
   backoffMs = 100,
 ) {
   let totalWait = 0;